requirements.txt
redis
python-dotenv
google-genai
numpy
scipy
ripser
cython
